simplification pass 1
