+++ apps/extension/package.json
@@ -28,7 +28,6 @@
     "dompurify": "^3.4.12",
     "fuse.js": "^7.5.0",
     "highlight.js": "^11.11.1",
-    "highlightjs-zig": "^1.0.2",
     "react": "^19.2.8",
     "react-dom": "^19.2.8",
     "zustand": "^5.0.14"
