+++ apps/extension/src/content/overlay/Overlay.tsx
@@ -1,11 +1,9 @@
 import { useCallback, useEffect, useId, useMemo, useRef, useState } from "react";
 import { useReviewStore, persistSession } from "./store";
-import { resolveUnitFiles } from "./selectors";
 import { buildDisplayUnits, displayUnitCount } from "./displayUnits";
-import { buildSelectableLines } from "./buildSelectableLines";
+import { buildSelectableLines, resolveUnitFiles } from "./buildSelectableLines";
 import { getFocusableElements, restoreFocusAfterOverlay } from "./focusTrap";
-import type { ViewChordPending } from "./viewModeChord";
-import { useOverlayKeyboard } from "./useOverlayKeyboard";
+import { useOverlayKeyboard, type ViewChordPending } from "./useOverlayKeyboard";
 import { useSubmitReviewFlow } from "./useSubmitReviewFlow";
 import { findUnitForFile, type DiffSearchResult, type SearchScrollTarget } from "./diffSearch";
 import { ProgressHeader } from "./components/ProgressHeader";
@@ -19,7 +17,7 @@ import { ConnectGitHubModal } from "./components/ConnectGitHubModal";
 import { confirm, ConfirmationHost, useConfirmationOpen } from "./components/confirmation";
 import { SubmitReviewModal } from "./components/SubmitReviewModal";
 import { ReviewSubmittedModal } from "./components/ReviewSubmittedModal";
-import { BUILD_PLAN_PRIMARY, buildPhaseDetail } from "./buildPhaseCopy";
+import { BUILD_PLAN_PRIMARY, buildPhaseDetail } from "./overlayCopy";
 
 interface OverlayProps {
   /** Invoked when the user exits so any in-flight stream can be cancelled. */
