--- apps/extension/src/content/overlay/diffSearch/index.ts
@@ -1,7 +1,0 @@
-export { buildDiffSearchIndex } from "./buildDiffSearchIndex";
-export { searchDiff } from "./searchDiff";
-export { findUnitForFile } from "./findUnitForFile";
-export { buildLinePreview } from "./linePreview";
-export type { PreviewLine } from "./linePreview";
-export { highlightSegments, fallbackMatchRanges } from "./highlightMatch";
-export type { DiffSearchDoc, DiffSearchResult, SearchScrollTarget, MatchRange } from "./types";
