--- apps/web/lib/analytics/index.ts
@@ -1,5 +1,0 @@
-export { getAnalyticsConfig } from "./config";
-export { createAnalyticsClient } from "./createClient";
-export { AnalyticsEvents, type CtaEventProperties, type CtaLocation } from "./events";
-export { createNoopClient } from "./noop";
-export type { AnalyticsClient, AnalyticsConfig } from "./types";
