--- apps/web/lib/analytics/noop.ts
@@ -1,9 +1,0 @@
-import type { AnalyticsClient } from "./types";
-
-export function createNoopClient(): AnalyticsClient {
-  return {
-    init() {},
-    capturePageview() {},
-    capture() {},
-  };
-}
